requirements.txt
pydantic
pydantic-settings
streamlit
pytest
pytest-asyncio
pytest-cov
google-cloud-bigquery
google-cloud-pubsub
google-cloud-aiplatform
vertexai
python-dotenv
google-generativeai
fastapi
uvicorn
requests
google-cloud-logging
google-cloud-storage
black
flake8